+++ samples/tab-deeplink/csharp/DeepLinkBot/AdapterWithErrorHandler.cs
