+++ samples/tab-deeplink/csharp/DeepLinkBot/DeeplinkHelper.cs
