+++ samples/tab-deeplink/csharp/DeepLinkBot/Manifest/color.png
