+++ samples/tab-deeplink/csharp/DeepLinkBot/Manifest/outline.png
