+++ samples/tab-deeplink/csharp/DeepLinkBot/Models/DeepLinkChannelModel.cs
