+++ samples/tab-deeplink/csharp/DeepLinkBot/Models/DeepLinksModel.cs
