+++ samples/tab-deeplink/csharp/DeepLinkBot/Views/Home/ChannelView.cshtml
