+++ samples/tab-deeplink/csharp/DeepLinkBot/Views/Home/Configure.cshtml
