+++ samples/tab-deeplink/csharp/DeepLinkBot/Views/Home/DeepLinkChannel.cshtml
