+++ samples/tab-deeplink/csharp/DeepLinkBot/Views/Home/Detail.cshtml
