+++ samples/tab-deeplink/csharp/DeepLinkBot/Views/Home/DetailChannel.cshtml
