+++ samples/tab-deeplink/csharp/DeepLinkBot/Views/Shared/_ValidationScriptsPartial.cshtml
